requirements.txt
aiofiles==24.1.0
aiogram==3.22.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aioschedule==0.5.2
aiosignal==1.4.0
annotated-types==0.7.0
APScheduler==3.11.0
arrow==1.3.0
attrs==25.3.0
certifi==2025.8.3
cffi==1.17.1
circuitbreaker==2.1.3
click==8.0.4
cryptography==44.0.3
frozenlist==1.7.0
idna==3.10
jmespath==0.10.0
magic-filter==1.0.12
multidict==6.6.4
oci==2.159.0
oci-cli==3.65.0
prompt-toolkit==3.0.43
propcache==0.3.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydub==0.25.1
pyOpenSSL==24.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
six==1.17.0
terminaltables==3.1.10
types-python-dateutil==2.9.0.20250822
typing-inspection==0.4.1
typing_extensions==4.15.0
tzlocal==5.3.1
wcwidth==0.2.13
yarl==1.20.1
